requirements-dev.txt
asgiref==3.4.1
Django==3.2.13
django-environ==0.8.1
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
